requirements.txt
gradio_imageslider
gradio==4.29.0
matplotlib
opencv-python
torch
torchvision
streamlit
streamlit-drawable-canvas
opencv-python-headless
numpy
Pillow
fpdf
plotly
rasterio
scikit-image
fpdf2
pyvista
stpyvista
scipy
streamlit-elements
scikit-learn
open3d